+++ src/test/resources/inspections/deadCode/UnusedUseVariableReferenceInspection.fixed.php
@@ -9,3 +9,15 @@ $dummy = function () use ($a, &$b, $c) {
 
     return $a + $c;
 };
+
+// Not applicable:
+
+$dummy = function () use (&$a) {
+    $a[] = 123;
+};
+
+$dummy = function () use (&$a) {
+    (function () use (&$a) {
+        $a = true;
+    })();
+};